chore: merge develop and adr adj
